requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.103.0
google-generativeai==0.3.0
clerk-sdk-python==0.1.0
APScheduler==3.10.4
python-dateutil==2.8.2